# force computation in opmath_t for CUDA fused optimizers