Fixes #153649

Benchmarks before and after:
```
(On RTX 5070)

Adam
=====================================

Main:
Mean time per run: 11066.4259 µs
Median: 11065.9903

With forced opmath_t:
Mean time per run: 8368.6145 µs
Median: 8367.6631

SGD
=====================================

Main:
Mean time per run: 3679.6347 µs
Median: 3678.9713

With forced opmath_t:
Mean time per run: 3518.2261 µs
Median: 3517.4846
```