@MeetThePatel 2025-05-27
```
Mismatched elements: 31 / 64 (48.4%)
Greatest absolute difference: 0.00017547607421875 at index (7, 0) (up to 3e-05 allowed)
Greatest relative difference: 1.2930149750900455e-05 at index (1, 7) (up to 1.3e-06 allowed)

To execute this test, run the following from the base repo dir:
    python test/test_cuda.py TestCudaOptimsCUDA.test_grad_scaling_autocast_fused_optimizers_AdamW_cuda_float32
```

I'm not as familiar with the AMP codebase, but are there any methods to mitigate the loss of precision from moving to `opmath_t` for the optimizer computations?